Refactor layouts and dialogs, add login loading state
Refactored layout components by creating a new `LayoutBase` class to centralize common properties and methods, and updated `LoginLayout` and `MainLayout` to inherit from it. Simplified version retrieval logic and removed redundant code.

Enhanced dialogs by adding `DefaultFocus` attributes and removing `MudFocusTrap` components. Improved `Login.razor` with a loading state to show a progress indicator during login. Added `CloseOnEscapeKey` to `MudDialogProvider` and initialized `CloseNestedDisabled` in `DiscoverDeviceDialog.razor.cs`.
