+++ Askstatus.Web.App/Layout/MainLayout.razor.cs
@@ -1,21 +1,14 @@
-﻿using System.Reflection;
-using System.Runtime.Versioning;
-using Microsoft.AspNetCore.Components;
-using MudBlazor;
+﻿using MudBlazor;
 
 namespace Askstatus.Web.App.Layout;
 
-public partial class MainLayout : LayoutComponentBase
+public partial class MainLayout : LayoutBase
 {
     private bool _isDarkMode;
 
     private bool _drawerOpen = false;
 
     private MudThemeProvider? _mudThemeProvider;
-
-    protected string Version { get; private set; } = string.Empty;
-
-    protected string AspDotnetVersion { get; private set; } = string.Empty;
 
     private void DrawerToggle()
     {
@@ -33,18 +26,5 @@ public partial class MainLayout : LayoutComponentBase
         {
             _isDarkMode = await _mudThemeProvider!.GetSystemPreference();
         }
-    }
-
-    protected override async Task OnInitializedAsync()
-    {
-        Assembly currentAssembly = typeof(MainLayout).Assembly;
-        if (currentAssembly == null)
-        {
-            currentAssembly = Assembly.GetCallingAssembly();
-        }
-
-        AspDotnetVersion = currentAssembly.GetCustomAttribute<TargetFrameworkAttribute>()?.FrameworkName!;
-        Version = $"{currentAssembly.GetName().Version!.Major}.{currentAssembly.GetName().Version!.Minor}.{currentAssembly.GetName().Version!.Build}";
-        await base.OnInitializedAsync();
     }
 }
