+++ Askstatus.Web.App/Pages/ChangePasswordDialog.razor
@@ -1,4 +1,4 @@
-﻿<MudDialog>
+﻿<MudDialog DefaultFocus="DefaultFocus">
     <TitleContent>
         <MudText Typo="Typo.h6">
             <MudIcon Icon="@Icons.Material.Filled.Security" Class="mr-3 mb-n1" />
@@ -7,11 +7,9 @@
     </TitleContent>
     <DialogContent>
         <MudForm Model="@ChangePassword" @ref="@Form" Validation="@ChangePasswordRequestValidator.ValidateValue" ValidationDelay="0">
-            <MudFocusTrap DefaultFocus="DefaultFocus">
-                <MudTextField InputType="InputType.Password" @bind-Value="@ChangePassword.OldPassword" Label="Old password" For="@(() => ChangePassword.OldPassword)" />
-                <MudTextField InputType="InputType.Password" @bind-Value="@ChangePassword.NewPassword" Label="New password" For="@(() => ChangePassword.NewPassword)" />
-                <MudTextField InputType="InputType.Password" @bind-Value="@ChangePassword.ConfirmPassword" Label="Confirm password" For="@(() => ChangePassword.ConfirmPassword)" />
-            </MudFocusTrap>
+            <MudTextField InputType="InputType.Password" @bind-Value="@ChangePassword.OldPassword" Label="Old password" For="@(() => ChangePassword.OldPassword)" />
+            <MudTextField InputType="InputType.Password" @bind-Value="@ChangePassword.NewPassword" Label="New password" For="@(() => ChangePassword.NewPassword)" />
+            <MudTextField InputType="InputType.Password" @bind-Value="@ChangePassword.ConfirmPassword" Label="Confirm password" For="@(() => ChangePassword.ConfirmPassword)" />
         </MudForm>
     </DialogContent>
     <DialogActions>
