+++ Askstatus.Web.App/Pages/Devices/DiscoverDeviceDialog.razor
@@ -22,16 +22,14 @@
             <MudText>@Device.Channel</MudText>
         }
         @*Nested dialog*@
-        <MudDialog @bind-Visible="nestedVisible">
+        <MudDialog @bind-Visible="nestedVisible" DefaultFocus="DefaultFocus">
             <TitleContent>
                 <MudText Typo="Typo.h6">
                     Enter device host/ip address!
                 </MudText>
             </TitleContent>
             <DialogContent>
-                <MudFocusTrap DefaultFocus="DefaultFocus">
-                    <MudTextField T="string" AutoFocus="true" DebounceInterval="100" OnDebounceIntervalElapsed="HandleIntervalElapsed" Variant="Variant.Filled" @bind-Value="@Url" Label="host/ip" />
-                </MudFocusTrap>
+                <MudTextField T="string" AutoFocus="true" DebounceInterval="100" OnDebounceIntervalElapsed="HandleIntervalElapsed" Variant="Variant.Filled" @bind-Value="@Url" Label="host/ip" />
             </DialogContent>
             <DialogActions>
                 <MudButton OnClick="CancelNested">Cancel</MudButton>
