+++ Askstatus.Web.App/Pages/Devices/EditDeviceDialog.razor
@@ -1,4 +1,4 @@
-﻿<MudDialog>
+﻿<MudDialog DefaultFocus="DefaultFocus">
     <TitleContent>
         <MudText Typo="Typo.h6">
             <MudIcon Icon="@Icons.Material.Filled.DeviceUnknown" Class="mr-3 mb-n1" />
@@ -6,18 +6,16 @@
         </MudText>
     </TitleContent>
     <DialogContent>
-        <MudFocusTrap DefaultFocus="DefaultFocus">
-            <MudTextField @bind-Value="@device.Name" Label="Name" />
-            <MudTextField @bind-Value="@device.DeviceId" Label="Id" />
-            <MudSelect T="Askstatus.Common.PowerDevice.ChanelType" @bind-Value="device.ChanelType" Label="Channel type">
-                @foreach (Askstatus.Common.PowerDevice.ChanelType item in Enum.GetValues(typeof(Askstatus.Common.PowerDevice.ChanelType)))
-                {
-                    <MudSelectItem T="Askstatus.Common.PowerDevice.ChanelType" Value="@item">
-                        <MudIcon Icon="@ChanelTypeToIcon(item)"></MudIcon> @item.ToString()
-                    </MudSelectItem>
-                }
-            </MudSelect>
-        </MudFocusTrap>
+        <MudTextField @bind-Value="@device.Name" Label="Name" />
+        <MudTextField @bind-Value="@device.DeviceId" Label="Id" />
+        <MudSelect T="Askstatus.Common.PowerDevice.ChanelType" @bind-Value="device.ChanelType" Label="Channel type">
+            @foreach (Askstatus.Common.PowerDevice.ChanelType item in Enum.GetValues(typeof(Askstatus.Common.PowerDevice.ChanelType)))
+            {
+                <MudSelectItem T="Askstatus.Common.PowerDevice.ChanelType" Value="@item">
+                    <MudIcon Icon="@ChanelTypeToIcon(item)"></MudIcon> @item.ToString()
+                </MudSelectItem>
+            }
+        </MudSelect>
     </DialogContent>
     <DialogActions>
         <MudButton OnClick="Cancel">Cancel</MudButton>
