+++ Askstatus.Web.App/Pages/Identity/ForgotPasswordDialog.razor
@@ -1,4 +1,4 @@
-﻿<MudDialog>
+﻿<MudDialog DefaultFocus="DefaultFocus">
     <TitleContent>
         <MudText Typo="Typo.h6">
             <MudIcon Icon="@Icons.Material.Filled.Security" Class="mr-3 mb-n1" />
@@ -7,9 +7,7 @@
     </TitleContent>
     <DialogContent>
         <MudForm Model="@ForgotPasswordRquest" @ref="@Form" Validation="@ForgotPasswordRquestValidator.ValidateValue" ValidationDelay="0">
-            <MudFocusTrap DefaultFocus="DefaultFocus">
-                <MudTextField InputType="InputType.Email" @bind-Value="@ForgotPasswordRquest.Email" Label="Email" For="@(() => ForgotPasswordRquest.Email)" />
-            </MudFocusTrap>
+            <MudTextField InputType="InputType.Email" @bind-Value="@ForgotPasswordRquest.Email" Label="Email" For="@(() => ForgotPasswordRquest.Email)" />
         </MudForm>
     </DialogContent>
     <DialogActions>
