+++ Askstatus.Web.App/Pages/Identity/Login.razor
@@ -27,33 +27,42 @@
     </Authorized>
     <NotAuthorized>
         <div style="height: 75vh; padding-top: 25vh">
-            <EditForm Model="@_model" OnValidSubmit="OnValidSubmit" Context="loginform">
-                <DataAnnotationsValidator />
-                <MudGrid Justify="Justify.Center">
-                    <MudItem xs="12" sm="3">
-                        <MudCard style="background-color: transparent; box-shadow: none;">
-                            <MudCardHeader Style="z-index: 1">
-                                <CardHeaderContent>
-                                    <MudText Typo="Typo.h6">Log in</MudText>
-                                    <ValidationSummary />
-                                </CardHeaderContent>
-                            </MudCardHeader>
-                            <MudCardContent Style="z-index: 1">
-                                <MudFocusTrap DefaultFocus="DefaultFocus">
-                                    <MudTextField Label="User name" HelperText="Enter user name" Variant="Variant.Filled"
-                                                  @bind-Value="_model.Username" For="@(() => _model.Username)" Adornment="Adornment.End" AdornmentIcon="@Icons.Material.Filled.Person" />
-                                    <MudTextField Label="Password" HelperText="Enter password" Class="mt-3" Variant="Variant.Filled"
-                                                  @bind-Value="_model.Password" For="@(() => _model.Password)" InputType="InputType.Password" Adornment="Adornment.End" AdornmentIcon="@Icons.Material.Filled.Password" />
-                                    <MudLink Underline="Underline.None" OnClick="ForgotPassword">Forgot password</MudLink>
-                                </MudFocusTrap>
-                            </MudCardContent>
-                            <MudCardActions>
-                                <MudButton ButtonType="ButtonType.Submit" Variant="Variant.Filled" Color="Color.Primary" Class="ml-auto">Log in</MudButton>
-                            </MudCardActions>
-                        </MudCard>
-                    </MudItem>
-                </MudGrid>
-            </EditForm>
+            @if (_loggingIn)
+            {
+            <MudGrid Justify="Justify.Center">
+                <MudProgressCircular Size="Size.Large" Indeterminate="true" />
+            </MudGrid>
+            }
+            else 
+            { 
+                <EditForm Model="@_model" OnValidSubmit="OnValidSubmit" Context="loginform">
+                    <DataAnnotationsValidator />
+                    <MudGrid Justify="Justify.Center">
+                        <MudItem xs="12" sm="3">
+                            <MudCard style="background-color: transparent; box-shadow: none;">
+                                <MudCardHeader Style="z-index: 1">
+                                    <CardHeaderContent>
+                                        <MudText Typo="Typo.h6">Log in</MudText>
+                                        <ValidationSummary />
+                                    </CardHeaderContent>
+                                </MudCardHeader>
+                                <MudCardContent Style="z-index: 1">
+                                    <MudFocusTrap DefaultFocus="DefaultFocus">
+                                        <MudTextField Label="User name" HelperText="Enter user name" Variant="Variant.Filled"
+                                                      @bind-Value="_model.Username" For="@(() => _model.Username)" Adornment="Adornment.End" AdornmentIcon="@Icons.Material.Filled.Person" />
+                                        <MudTextField Label="Password" HelperText="Enter password" Class="mt-3" Variant="Variant.Filled"
+                                                      @bind-Value="_model.Password" For="@(() => _model.Password)" InputType="InputType.Password" Adornment="Adornment.End" AdornmentIcon="@Icons.Material.Filled.Password" />
+                                        <MudLink Underline="Underline.None" OnClick="ForgotPassword">Forgot password</MudLink>
+                                    </MudFocusTrap>
+                                </MudCardContent>
+                                <MudCardActions>
+                                    <MudButton ButtonType="ButtonType.Submit" Variant="Variant.Filled" Color="Color.Primary" Class="ml-auto">Log in</MudButton>
+                                </MudCardActions>
+                            </MudCard>
+                        </MudItem>
+                    </MudGrid>
+                </EditForm>
+            }
         </div>
     </NotAuthorized>
 </AuthorizeView>
