+++ Askstatus.Web.App/Pages/Identity/ResetPasswordDialog.razor
@@ -1,4 +1,4 @@
-﻿<MudDialog>
+﻿<MudDialog DefaultFocus="DefaultFocus">
     <TitleContent>
         <MudText Typo="Typo.h6">
             <MudIcon Icon="@Icons.Material.Filled.Security" Class="mr-3 mb-n1" />
@@ -7,9 +7,7 @@
     </TitleContent>
     <DialogContent>
         <MudForm Model="@ResetPasswordRequest" @ref="@Form" Validation="@ResetPasswordRequestFluentValidator.ValidateValue" ValidationDelay="0">
-            <MudFocusTrap DefaultFocus="DefaultFocus">
-                <MudTextField InputType="InputType.Password" @bind-Value="@ResetPasswordRequest.NewPassword" Label="Password" For="@(() => ResetPasswordRequest.NewPassword)" />
-            </MudFocusTrap>
+            <MudTextField InputType="InputType.Password" @bind-Value="@ResetPasswordRequest.NewPassword" Label="Password" For="@(() => ResetPasswordRequest.NewPassword)" />
         </MudForm>
     </DialogContent>
     <DialogActions>
